requirements.txt
pyspark==3.4.1
langchain==0.1.0
openai>=1.6.1,<2.0.0
langchain-openai==0.0.2
instructor==0.4.5
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.24.3
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1
pandas==2.1.4
langgraph==0.0.18
faker==20.1.0
plotly==5.18.0 